requirements.txt
aiogram==3.22.0
openai==2.8.1
python-dotenv==1.2.1
asyncpg==0.31.0
tiktoken==0.12.0
apscheduler==3.11.1